moved shared utils
