+++ src/app/content/highlights/components/Card.spec.tsx
@@ -14,8 +14,9 @@ import { assertDocument } from '../../../utils';
 import { openToc, receiveBook, receivePage } from '../../actions';
 import { highlightStyles } from '../../constants';
 import { requestSearch } from '../../search/actions';
+import { highlightLocationFilters } from '../../selectors';
+import { HighlightData } from '../../types';
 import { formatBookData } from '../../utils';
-import { getHighlightLocationFilterForPage } from '../../utils/sharedHighlightsUtils';
 import {
   createHighlight,
   deleteHighlight,
@@ -23,8 +24,7 @@ import {
   receiveHighlights,
   setAnnotationChangesPending,
 } from '../actions';
-import { highlightLocationFilters } from '../selectors';
-import { HighlightData } from '../types';
+import { getHighlightLocationFilterForPage } from '../utils';
 import Card, { CardProps } from './Card';
 import DisplayNote from './DisplayNote';
 import EditCard from './EditCard';
